refactor(engine): decompose delete/context.rs into per-concern submodules
Splits the 1020-LoC `delete/context.rs` into focused submodules under
`delete/context/` while preserving the public API exactly:

- `core.rs`  - `DeleteContext` struct, constructors, observation API,
               and `emit_*`/`into_emitter` drain plumbing.
- `outcome.rs` - `DrainOutcome` plus its `from_emitter` constructor.
- `timing.rs` - `EmitterTiming` enum, its predicates, and `From`
                conversions to/from `local_copy::DeleteTiming`.
- `tests.rs`  - the existing unit tests, unchanged.
- `mod.rs`    - module documentation and re-exports of the public API.

`DeleteContext::new` cursor seating (`cursor_root = dest_root.clone()`)
is preserved verbatim. No behavioural changes; `cargo check --workspace
--all-features` and the `--tests` target both build clean.
